feat: add supabase error codes
